requirements.txt
streamlit>=1.32.0
openai>=2.0.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
python-docx>=1.1.0
numpy>=1.24.0
